% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{all_interpolate}
\alias{all_interpolate}
\title{Interpolating points to data of list}
\usage{
all_interpolate(collect, nvic)
}
\arguments{
\item{collect}{data to be interpolated}

\item{nvic}{the number of neighbor points}
}
\value{
list of interpolated data sets
}
\description{
Interpolating points to data of list
}
