% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{convex_hull_check}
\alias{convex_hull_check}
\title{Distinguishing whether interpolated points are in a convex made of projected neighbor points via PCA from origin data.}
\usage{
convex_hull_check(rpca, hline, sides)
}
\arguments{
\item{rpca}{projected neighbor points via PCA}

\item{hline}{horizontal lines from interpolated points}

\item{sides}{a side set of a convex}
}
\value{
True values indicate whether horizontal lines from interpolated points cross sides of convex in the side set
}
\description{
Distinguishing whether interpolated points are in a convex made of projected neighbor points via PCA from origin data.
}
