% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{convex_hull_vertx}
\alias{convex_hull_vertx}
\title{Deriving sides having a specified vertex of a convex}
\usage{
convex_hull_vertx(chul, vertx)
}
\arguments{
\item{chul}{vertex set of a convex}

\item{vertx}{a vertex to derive what sides have this vertex}
}
\value{
side set having "vertx"
}
\description{
Deriving sides having a specified vertex of a convex
}
