% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{exist_convexhull_check}
\alias{exist_convexhull_check}
\title{Distinguishing whether passed points are in convex.}
\usage{
exist_convexhull_check(rpca, insecs)
}
\arguments{
\item{rpca}{points consisting a convex}

\item{insecs}{points to be distinguished}
}
\value{
True when points are in convex
}
\description{
Distinguishing whether passed points are in convex.
}
