% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{get_vicinity}
\alias{get_vicinity}
\title{Choose 1-n th close points to a point}
\usage{
get_vicinity(dist, center, nvic)
}
\arguments{
\item{dist}{distant matrix}

\item{center}{a vector element number of center of the neighborhoods}

\item{nvic}{the number of neighborhood points}
}
\value{
vector element numbers the neighborhoods
}
\description{
Choose 1-n th close points to a point
}
