% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{origin_coordinate}
\alias{origin_coordinate}
\title{Projecting intepolated points to coordinate system of original data}
\usage{
origin_coordinate(rpca, incord, ori_cen)
}
\arguments{
\item{rpca}{points whose dimension are reduced via PCA}

\item{incord}{interpolated points}

\item{ori_cen}{a center point of neighbors}
}
\value{
coordinates of intepolated points in coordinate system of original data
}
\description{
Projecting intepolated points to coordinate system of original data
}
