% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{sides_set}
\alias{sides_set}
\title{Delete the same side in side set}
\usage{
sides_set(sides)
}
\arguments{
\item{sides}{a side set}
}
\value{
a side set without overlapb
}
\description{
Delete the same side in side set
}
