% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{voronoi_border}
\alias{voronoi_border}
\title{Add points at the vertexes of Voronoi region}
\usage{
voronoi_border(vics, figure)
}
\arguments{
\item{vics}{center and neighbor points}

\item{figure}{data set to be interpolated}
}
\value{
added points
}
\description{
Add points at the vertexes of Voronoi region
}
