% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{voronoi_interpo}
\alias{voronoi_interpo}
\title{Interpolating points to one data set}
\usage{
voronoi_interpo(figure, nvics)
}
\arguments{
\item{figure}{data set}

\item{nvics}{the number of neighborhood points}
}
\description{
Interpolating points to one data set
}
