% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolation.R
\name{all_interpolate}
\alias{all_interpolate}
\title{Interpolating points to data of list}
\usage{
all_interpolate(collect)
}
\arguments{
\item{collect}{data to be interpolated}
}
\description{
Interpolating points to data of list
}
